+++ lib/widgets/conversations/messages_page/message_entry/message_entry.dart
@@ -1,5 +1,5 @@
-import 'package:crowdleague/widgets/conversations/conversation_page/message_entry/message_text_field.dart';
-import 'package:crowdleague/widgets/conversations/conversation_page/message_entry/save_message_button.dart';
+import 'package:crowdleague/widgets/conversations/messages_page/message_entry/message_text_field.dart';
+import 'package:crowdleague/widgets/conversations/messages_page/message_entry/save_message_button.dart';
 import 'package:flutter/material.dart';
 
 class MessageEntry extends StatelessWidget {
