+++ lib/widgets/conversations/messages_page/messages_list.dart
@@ -1,7 +1,7 @@
 import 'package:built_collection/built_collection.dart';
 import 'package:crowdleague/models/app/app_state.dart';
 import 'package:crowdleague/models/conversations/conversation/message.dart';
-import 'package:crowdleague/widgets/conversations/conversation_page/message_tile.dart';
+import 'package:crowdleague/widgets/conversations/messages_page/message_tile.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
