+++ lib/widgets/conversations/messages_page/messages_page.dart
@@ -1,8 +1,8 @@
 import 'package:crowdleague/actions/conversations/disregard_messages.dart';
 import 'package:crowdleague/actions/conversations/observe_messages.dart';
 import 'package:crowdleague/models/app/app_state.dart';
-import 'package:crowdleague/widgets/conversations/conversation_page/message_entry/message_entry.dart';
-import 'package:crowdleague/widgets/conversations/conversation_page/messages_list.dart';
+import 'package:crowdleague/widgets/conversations/messages_page/message_entry/message_entry.dart';
+import 'package:crowdleague/widgets/conversations/messages_page/messages_list.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
